Move heartbeat subscription to service and fix interval so that it executes on the leading edge
