+++ constants/constants.go
@@ -1,6 +1,7 @@
 package constants
 
 import (
+	"errors"
 	"github.com/ethereum/go-ethereum/common"
 )
 
@@ -109,3 +110,6 @@ var GanacheAccountToPrivateKey = map[common.Address][]byte{
 	GanacheAccount3: ganacheAccount3PrivateKey,
 	GanacheAccount4: ganacheAccount4PrivateKey,
 }
+
+// ErrInternal is used whenever we don't wish to expose internal errors to a client
+var ErrInternal = errors.New("internal error")
